+++ frontend/main/src/pages/main/Auth/components/Login/index.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { LoginForm } from './LoginForm';
 import { useNavigate } from 'react-router-dom';
 import Loading from '@/components/Loading';
-import { useLogin } from '@/hooks/auth';
+import { useLogin } from '@/hooks/login';
 import { useProfile } from '@/hooks/profile';
 export const Login = () => {
     const [loading, setLoading] = useState(false);
@@ -19,8 +19,12 @@ export const Login = () => {
     }, [dataLogin, profile]);
     useEffect(() => {
         if (dataProfile?.role === 'admin') {
+            localStorage.setItem('role', 'admin');
+            localStorage.setItem('profile', JSON.stringify(dataProfile));
             navigate('/admin');
         } else if (dataProfile?.role === 'user') {
+            localStorage.setItem('role', 'user');
+            localStorage.setItem('profile', JSON.stringify(dataProfile));
             navigate('/');
         }
     }, [dataProfile, navigate]);
