+++ frontend/main/src/service/authRegister.js
@@ -1,0 +1,7 @@
+import { GenaralService } from './general';
+
+export class AuthRegisterService extends GenaralService {
+    constructor() {
+        super('auth/register');
+    }
+}
