+++ frontend/main/src/service/container.js
@@ -7,6 +7,7 @@ import { ActorService } from './actor';
 import { WatchService } from './watch';
 import { AuthLoginService } from './authLogin';
 import { ProfileService } from './profile';
+import { AuthRegisterService } from './authRegister';
 export const appContainer = awilix.createContainer();
 
 appContainer.register({
@@ -18,4 +19,5 @@ appContainer.register({
     watchService: awilix.asClass(WatchService).singleton(),
     authLoginService: awilix.asClass(AuthLoginService).singleton(),
     profileService: awilix.asClass(ProfileService).singleton(),
+    authRegisterService: awilix.asClass(AuthRegisterService).singleton(),
 });
